+++ test/unit/fuzz-tools.test.js
@@ -230,6 +230,7 @@ import * as cardioechoV237 from '../../lib/cardioecho-v237.js';
 import * as anthroV238 from '../../lib/anthro-v238.js';
 import * as gisurgV239 from '../../lib/gisurg-v239.js';
 import * as rehabV240 from '../../lib/rehab-v240.js';
+import * as geriV241 from '../../lib/geri-v241.js';
 
 const MODULES = {
   'clinical.js': clinical,
@@ -404,6 +405,7 @@ const MODULES = {
   'anthro-v238.js': anthroV238,
   'gisurg-v239.js': gisurgV239,
   'rehab-v240.js': rehabV240,
+  'geri-v241.js': geriV241,
 };
 
 const MATRIX = [0, -1, 1e9, NaN, Infinity, -Infinity, '', undefined, null];